Chore: move utility functions to pkg/utils/k8s/k8s.go and fix new tests
